Update-0ff1b8f8-9ef7-44ec-92f3-c69889778f79

+++ src/main/java/com/bss/inc/cameraon/utils/net/angryscan/core/ScanningProgressCallback.java
@@ -14,5 +14,5 @@ public interface ScanningProgressCallback {
      * @param percentageComplete value from 0 to 100, showing how much work
      * 		is already done.
      */
-    void updateProgress(InetAddress currentAddress, int runningThreads, int percentageComplete);
+    void updateProgress(InetAddress currentAddress, int runningThreads, double percentageComplete);
 }
